PB/requirements.txt
django==4.1.2
Pillow==9.3.0
djangorestframework-simplejwt==5.2.2
pyjwt==2.6.0
django-cors-headers=3.13.0